+++ pkg/utils/validation.go
@@ -18,7 +18,7 @@ type ValidationError struct {
 // ValidateStruct validates a struct and returns formatted error messages
 func ValidateStruct(s interface{}) []ValidationError {
 	var validationErrors []ValidationError
-	
+
 	if err := validate.Struct(s); err != nil {
 		for _, err := range err.(validator.ValidationErrors) {
 			var message string
@@ -34,14 +34,14 @@ func ValidateStruct(s interface{}) []ValidationError {
 			default:
 				message = "Invalid value"
 			}
-			
+
 			validationErrors = append(validationErrors, ValidationError{
 				Field:   ToSnakeCase(err.Field()),
 				Message: message,
 			})
 		}
 	}
-	
+
 	return validationErrors
 }
 
@@ -50,17 +50,17 @@ func ValidateWalletUserID(walletUserID string) error {
 	if walletUserID == "" {
 		return NewWalletError(CodeValidationError, "wallet_user_id is required", "")
 	}
-	
+
 	if len(walletUserID) > 255 {
 		return NewWalletError(CodeValidationError, "wallet_user_id is too long", "maximum 255 characters allowed")
 	}
-	
+
 	// Basic validation - alphanumeric, hyphens, underscores
 	matched, _ := regexp.MatchString(`^[a-zA-Z0-9_-]+$`, walletUserID)
 	if !matched {
 		return NewWalletError(CodeValidationError, "invalid wallet_user_id format", "only alphanumeric characters, hyphens, and underscores are allowed")
 	}
-	
+
 	return nil
 }
 
@@ -70,26 +70,27 @@ func ValidateBalanceType(balanceType string) error {
 		"coins": true,
 		"exp":   true,
 	}
-	
+
 	if !validTypes[balanceType] {
 		return NewWalletError(CodeInvalidBalanceType, "invalid balance type", "allowed types: coins, exp")
 	}
-	
+
 	return nil
 }
 
-// ValidateAmount validates amount value
-func ValidateAmount(amount int64) error {
+// ValidateAmount validates amount value (float64)
+func ValidateAmount(amount float64) error {
 	if amount <= 0 {
 		return NewWalletError(CodeInvalidAmount, "amount must be positive", "")
 	}
-	
+
 	if amount > 1000000000 { // 1 billion limit
 		return NewWalletError(CodeInvalidAmount, "amount is too large", "maximum 1,000,000,000 allowed")
 	}
-	
+
 	return nil
 }
+
 
 // ToSnakeCase converts camelCase to snake_case
 func ToSnakeCase(str string) string {
